+++ python/pyrogue/_Device.py
@@ -663,82 +663,3 @@ class DataWriter(Device):
 
         self.dataFile.set(base + datetime.datetime.now().strftime("data_%Y%m%d_%H%M%S.dat")) 
 
-class RunControl(Device):
-    """Special base class to control runs. TODO: Update comments."""
-
-    def __init__(self, *, hidden=True, rates=None, states=None, cmd=None, **kwargs):
-        """Initialize device class"""
-
-        if rates is None:
-            rates={1:'1 Hz', 10:'10 Hz'}
-
-        if states is None:
-            states={0:'Stopped', 1:'Running'}
-
-        Device.__init__(self, hidden=hidden, **kwargs)
-
-        value = [k for k,v in states.items()][0]
-
-        self._thread = None
-        self._cmd = cmd
-
-        self.add(pr.LocalVariable(
-            name='runState',
-            value=value,
-            mode='RW',
-            disp=states,
-            localSet=self._setRunState,
-            description='Run state of the system.'))
-
-        value = [k for k,v in rates.items()][0]
-
-        self.add(pr.LocalVariable(
-            name='runRate',
-            value=value,
-            mode='RW',
-            disp=rates,
-            localSet=self._setRunRate,
-            description='Run rate of the system.'))
-
-        self.add(pr.LocalVariable(
-            name='runCount',
-            value=0,
-            typeStr='UInt32',
-            mode='RO',
-            pollInterval=1,
-            description='Run Counter updated by run thread.'))
-
-    def _setRunState(self,value,changed):
-        """
-        Set run state. Reimplement in sub-class.
-        Enum of run states can also be overriden.
-        Underlying run control must update runCount variable.
-        """
-        if changed:
-            if self.runState.valueDisp() == 'Running':
-                #print("Starting run")
-                self._thread = threading.Thread(target=self._run)
-                self._thread.start()
-            elif self._thread is not None:
-                #print("Stopping run")
-                self._thread.join()
-                self._thread = None
-
-    def _setRunRate(self,value):
-        """
-        Set run rate. Reimplement in sub-class if neccessary.
-        """
-        pass
-
-    def _run(self):
-        #print("Thread start")
-        self.runCount.set(0)
-
-        while (self.runState.valueDisp() == 'Running'):
-            time.sleep(1.0 / float(self.runRate.value()))
-            if self._cmd is not None:
-                self._cmd()
-
-            self.runCount.set(self.runCount.value() + 1,write=False)
-        #print("Thread stop")
-
