+++ python/pyrogue/__init__.py
@@ -24,6 +24,7 @@ from pyrogue._Memory    import *
 from pyrogue._Root      import *
 from pyrogue._PollQueue import *
 from pyrogue._Virtual   import *
+from pyrogue._Process   import *
 
 def addLibraryPath(path):
     """
